feat(gatekeeper): add new create checkout session function
