+++ packages/server/modules/gatekeeper/clients/getResultUrl.ts
@@ -1,0 +1,9 @@
+export const getResultUrl = ({
+  frontendOrigin,
+  workspaceId,
+  workspaceSlug
+}: {
+  frontendOrigin: string
+  workspaceSlug: string
+  workspaceId: string
+}) => new URL(`${frontendOrigin}/workspaces/${workspaceSlug}?workspace=${workspaceId}`)
